// src/generated/models/Fc_placesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Fc_placesfc_level = {
  122370000: 'country',
  122370001: 'region',
  122370002: 'city',
  122370003: 'neighbourhood'
} as const;
export type Fc_placesfc_level = keyof typeof Fc_placesfc_level;
export const Fc_placesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Fc_placesstatecode = keyof typeof Fc_placesstatecode;
export const Fc_placesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Fc_placesstatuscode = keyof typeof Fc_placesstatuscode;

export interface Fc_placesBase {
  fc_code: string;
  "fc_Company@odata.bind"?: string;
  fc_latitude?: number;
  fc_level: Fc_placesfc_level;
  fc_longitude?: number;
  fc_name?: string;
  "fc_Parent@odata.bind"?: string;
  fc_placeid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Fc_placesstatecode;
  statuscode?: Fc_placesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Fc_places extends Fc_placesBase {
  createdon?: string;
  fc_levelname?: string;
  modifiedon?: string;
  ownerid: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  fc_company?: object;
  _fc_company_value?: string;
  fc_parent?: object;
  _fc_parent_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}